# backend/app/static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "IntelliDesk AI — API Specification",
    "description": "Comprehensive interactive REST API documentation for the IntelliDesk AI platform, including user authentication, ticketing, SLA management, incident handling, and AI-powered knowledge base operations.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Local Gateway (default)"
    },
    {
      "url": "http://localhost:8000/api/v1",
      "description": "Flask Backend Direct"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/ai/chat": {
      "post": {
        "tags": ["AI Assistant"],
        "summary": "AI Chat Assistant",
        "description": "Send a chat message to the RAG-enabled AI support assistant.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["message"],
                "properties": {
                  "message": { "type": "string", "example": "How do I connect to VPN?" },
                  "session_uuid": { "type": "string", "example": "123e4567-e89b-12d3-a456-426614174000" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "AI assistant response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "success" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "response": { "type": "string", "example": "To connect to the VPN, make sure you download Cisco AnyConnect..." },
                        "session_uuid": { "type": "string" },
                        "sources": { "type": "array", "items": { "type": "object" } }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/ai/sessions/{uuid}": {
      "get": {
        "tags": ["AI Assistant"],
        "summary": "Get Session History",
        "parameters": [
          { "name": "uuid", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Messages history array" }
        }
      },
      "delete": {
        "tags": ["AI Assistant"],
        "summary": "Delete Chat Session",
        "parameters": [
          { "name": "uuid", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Session deleted" }
        }
      }
    },
    "/ai/tickets/{id}/classify": {
      "post": {
        "tags": ["AI Assistant"],
        "summary": "Classify Ticket",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "AI predicted category and severity" }
        }
      }
    },
    "/ai/tickets/{id}/classification": {
      "get": {
        "tags": ["AI Assistant"],
        "summary": "Get Classification Results",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Stored AI classification data" }
        }
      }
    },
    "/ai/tickets/{id}/classification/feedback": {
      "post": {
        "tags": ["AI Assistant"],
        "summary": "Provide Classification Feedback",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Feedback submitted successfully" }
        }
      }
    },
    "/ai/tickets/{id}/suggest-response": {
      "post": {
        "tags": ["AI Assistant"],
        "summary": "Suggest Response Draft",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "AI drafted agent reply" }
        }
      }
    },
    "/ai/tickets/{id}/summarize": {
      "post": {
        "tags": ["AI Assistant"],
        "summary": "Summarize Ticket Thread",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Text summary of comments" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Pass your JWT token in the Authorization header: `Bearer <token>`"
      }
    },
...
  }
}